Change ticket type

+++ feature/todo/src/main/java/jp/kztproject/rewardedtodo/todo/domain/EditingTodo.kt
@@ -4,7 +4,7 @@ data class EditingTodo(
     var id: Long? = null,
     val todoistId: Long? = null,
     var name: String = "",
-    private var numberOfTicketsObtained: Float = 0f,
+    private var numberOfTicketsObtained: Int = 0,
     private var isRepeat: Boolean = false
 ) {
